% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_array}
\alias{as_array}
\title{As Array}
\usage{
as_array(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
(any)
}
\description{
Convert a value (usually a tensor from some framework) to an R array.
}
