% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_dtype}
\alias{as_dtype}
\title{Convert to DataType}
\usage{
as_dtype(x)
}
\arguments{
\item{x}{(any)\cr
A dtype string (e.g. \code{"f32"}, \code{"i64"}, \code{"bool"}; the aliases \code{"pred"}
and \code{"i1"} map to \code{"bool"}) or a \code{\link{DataType}} object.}
}
\value{
\code{DataType}
}
\description{
Convert to \code{\link{DataType}}.
}
