% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{assert_dtype}
\alias{assert_dtype}
\title{Assert TensorDataType}
\usage{
assert_dtype(x, arg = rlang::caller_arg(x))
}
\arguments{
\item{x}{(any)\cr
Object to check.}

\item{arg}{(\code{character(1)})\cr
Argument name for error messages.}
}
\value{
\code{NULL} (invisibly). Called for side effects.
}
\description{
Assert that an object is a \code{\link{TensorDataType}}.
}
