% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ComplexType}
\alias{ComplexType}
\title{ComplexType}
\usage{
ComplexType(value)
}
\arguments{
\item{value}{(\code{integer(1)})\cr
Bit width of the underlying floating point component (\code{32} or \code{64}).}
}
\value{
\code{ComplexType}
}
\description{
Represents a complex floating point type. The \code{value} argument is the bit
width of the underlying real component (i.e. \code{32} for \verb{complex<f32>} /
\code{c64}, and \code{64} for \verb{complex<f64>} / \code{c128}).
}
