% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device}
\alias{device}
\title{Device}
\usage{
device(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
(any)
}
\description{
Get the device of a tensor.
}
