% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{dtype_bits}
\alias{dtype_bits}
\title{Bit Width of a DataType}
\usage{
dtype_bits(x)
}
\arguments{
\item{x}{(\code{\link{DataType}})}
}
\value{
\code{integer(1)}
}
\description{
The number of bits of one element of the given \code{\link{DataType}}.
For \code{bool} this is \code{1}; for complex types it is the total width of the
element (\code{c64} -> \code{64}).
}
