% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{dtype_category}
\alias{dtype_category}
\title{Category of a DataType}
\usage{
dtype_category(x)
}
\arguments{
\item{x}{(\code{\link{DataType}})}
}
\value{
\code{character(1)}
}
\description{
The category of a \code{\link{DataType}}: one of \code{"bool"}, \code{"int"}, \code{"uint"},
\code{"float"}, or \code{"complex"}.
}
\seealso{
\code{\link[=is_dtype_float]{is_dtype_float()}} and friends for per-category predicates
}
