% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FloatType}
\alias{FloatType}
\title{FloatType}
\usage{
FloatType(value)
}
\arguments{
\item{value}{(\code{integer(1)})}
}
\value{
\code{FloatType}
}
\description{
Represents a floating point type with a given bit width.
}
