% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{IntegerType}
\alias{IntegerType}
\title{IntegerType (signed)}
\usage{
IntegerType(value)
}
\arguments{
\item{value}{(\code{integer(1)})}
}
\value{
\code{IntegerType}
}
\description{
Represents a signed integer type with a given bit width.
}
