% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_dtype}
\alias{is_dtype}
\title{Is TensorDataType}
\usage{
is_dtype(x)
}
\arguments{
\item{x}{(any)\cr
Object to check.}
}
\value{
\code{logical(1)}
}
\description{
Check if an object is a \code{\link{TensorDataType}}.
}
