% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{shape}
\alias{shape}
\title{Shape}
\usage{
shape(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
(\code{integer()})\cr
The shape of the tensor.
}
\description{
Get the shape of a tensor.
}
