% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{UIntegerType}
\alias{UIntegerType}
\title{UIntegerType}
\usage{
UIntegerType(value)
}
\arguments{
\item{value}{(\code{integer(1)})}
}
\value{
\code{UIntegerType}
}
\description{
Represents an unsigned integer type with a given bit width.
}
