% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_dtype}
\alias{as_dtype}
\title{Convert to DataType}
\usage{
as_dtype(x)
}
\arguments{
\item{x}{(any)\cr
Object to convert.
Can currently be a string (e.g. \code{"i1"}, \code{"i32"}, \code{"f32"}, ...) or a \code{\link{DataType}} object.}
}
\value{
\code{DataType}
}
\description{
Convert to DataType.
}
